requirements.txt
cassandra-driver
pandas
numpy